[Fix] #2068 라벨 paint-order 게이트를 .vec draw order 대조로 재작성
page-background(전면 흰 rect) 반입으로 "흰 픽셀 = halo" 분류가 성립하지 않게 된
두 번째 픽셀 게이트를 배경 무관 판정으로 바꾸고, 배포 pack metadata의 드리프트를
정정한다.

- route_map_basemap_label_paint_order_test: 대표 라벨 6건 래스터 측정(맞춘 하한
  0.25~0.6)을 버리고 .vec 그리기 순서를 직접 본다 — 같은 글자·같은 앵커의
  stroke 전용 draw가 fill 전용 draw보다 먼저인지 5권역 전수. 실측 결과 순서 위반
  0건이고 halo/글자 짝은 seoul 0 · busan 155 · daegu 104 · daejeon 22 · gwangju 20.
  대전·광주 halo는 자손 결합자 CSS 규칙 유실로 빠져 있다가 이번에 복원된 것이다.
- test/support/basemap_vec_text_draws.dart: 두 게이트가 공유하는 .vec 텍스트 draw
  디코더(위치·fontSize·fill/stroke 구분)를 한 벌로 둔다.
- metro_map_pack manifest의 compiled_content를 전량 반입 실태로 정정
  (route-lines-and-station-symbols -> owner-svg-verbatim-full-document).

Refs #2068
